requirements.txt
streamlit
pandas
matplotlib
google-generativeai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client